Pre Upload to renaming to package to OPT. Cannot run before renmed to OPT
Immense changes to Docstrings. Moved Commands to Macros_Commands package, which was previously Macros package. Removed multiple functions in _Timeline_generator modules where it was unnecceassary. Updated duration calculations of certain modes to allow the runtime of CMDs to have enough time.  Tweaks to Timeline_Plotter to make it closer to the actually Commands generated and the data from OHB